serverless/functions/requirements.txt
firebase_functions~=0.1.0
google-auth
google-cloud-tasks